requirements.txt
fastapi
uvicorn
pydantic
aiosqlite
python-multipart
dotenv
openai
pypdf
requests
python-docx